style: 구조 분리 및 변경
