+++ src/css/app.css
@@ -45,7 +45,8 @@
   flex-direction: column;
   align-items: center;
   overflow: scroll;
-  height: 430px;
+  overflow: hidden;
+  height: 390px;
   margin: 10px 0;
 }
