+++ src/js/Observer.js
@@ -1,0 +1,17 @@
+class Observer {
+  constructor() {
+    this.observers = [];
+  }
+
+  subscribe(observer) {
+    this.observers.push(observer);
+  }
+
+  notify() {
+    this.observers.forEach(observer => observer.show());
+  }
+}
+
+const observer = new Observer();
+
+export { observer };
